Updated Main.ccp, created ObjectDetector.CPP and OutputProcessing_BM.cpp
Added 3 ways to detect objects 
Red: Given the size of the box find the position where most features are inside the box, it tries to lay the box both vertically and horizontally
Blue: Same as red but centered around the final features that can fit in the box
Green: Used canny edge detection together with find contours, find a box that fills the contour. From the list of boxes given from canny + contours finds the box witch contains the most features
I don't feel the results is very good, but to improve it I think we need better feature detection. But if you guys have any other idea to improve i'm all ears